model-build-train/seed_code/pipelines/foot_traffic/requirements/02_training_prep_reqs.txt
pyyaml
sagemaker
boto3
pandas
pyarrow
snowflake-connector-python[pandas]
snowflake-snowpark-python[pandas]